requirements.txt
torch
torchvision
pillow
matplotlib
numpy
opencv-python
open_clip_torch
